% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{cens_tabelle.lst}
\alias{cens_tabelle.lst}
\title{Lista delle tabelle}
\format{
An object of class \code{character} of length 1.
}
\usage{
cens_tabelle.lst
}
\description{
Lista delle tabelle
}
\keyword{datasets}
