% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{cens_tipi_zone.lst}
\alias{cens_tipi_zone.lst}
\title{Lista dei Tipi di Localita}
\format{
An object of class \code{character} of length 1.
}
\usage{
cens_tipi_zone.lst
}
\description{
Lista dei Tipi di Localita
}
\keyword{datasets}
