% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dati.R
\docType{data}
\name{eta05}
\alias{eta05}
\title{eta05}
\format{
Una data.table con i campi seguenti:
\describe{
  \item{\code{CMN}}{ Codice ISTAT del Comune (si veda la tabella \code{comuni} nel pacchetto \code{masteRgeo} ) }
  \item{\code{sesso}}{ \code{M}-Maschio, \code{F}-Femmina, \code{T}-Totale }
  \item{\code{eta}}{ Classi di eta quinquennali (si veda la tabella \code{cens_diz} per la descrizione delle etichette) }
  \item{\code{anno}}{ Anno di riferimento }
  \item{\code{valore}}{ Valore del Dato }
}
}
\usage{
eta05
}
\description{
Tabella contenente il conteggio della Popolazione residente raggruppata per Sesso ed eta in classi quinquennali
}
\details{
La tabella è indicizzata secondo i campi \code{CMN}, \code{sesso}, \code{eta}, \code{anno}.
}
\keyword{datasets}
