% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dati.R
\docType{data}
\name{eta10}
\alias{eta10}
\title{eta10}
\format{
Una data.table con i campi seguenti:
\describe{
  \item{\code{CMN}}{ Codice ISTAT del Comune (si veda la tabella \code{comuni} nel pacchetto \code{masteRgeo} ) }
  \item{\code{sesso}}{ \code{M}-Maschio, \code{F}-Femmina, \code{T}-Totale }
  \item{\code{cittadinanza}}{ \code{I}-Italiano, \code{S}-Straniero, \code{T}-Totale }
  \item{\code{eta}}{ Classi di eta decennali (si veda la tabella \code{cens_diz} per la descrizione delle etichette) }
  \item{\code{anno}}{ Anno di riferimento }
  \item{\code{valore}}{ Valore del Dato }
}
}
\usage{
eta10
}
\description{
Tabella contenente il conteggio della Popolazione residente raggruppata per Sesso,cittadinanza ed eta in classi decennali
}
\details{
La tabella è indicizzata secondo i campi \code{CMN}, \code{sesso}, \code{eta}, \code{cittadinanza}, \code{anno}.
}
\keyword{datasets}
