% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dati.R
\docType{data}
\name{etichette}
\alias{etichette}
\title{etichette}
\format{
Una data.table con i campi seguenti:
\describe{
  \item{\code{tabella}}{ Nome Tabella }
  \item{\code{attributo}}{ Nome Variabile }
  \item{\code{valore}}{ Valore }
  \item{\code{etichetta}}{ Etichetta }
}
}
\usage{
etichette
}
\description{
Descrizione dei Valori assunti dagli Attributi presenti nelle Tabelle
}
\keyword{datasets}
