% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dati.R
\docType{data}
\name{famiglie}
\alias{famiglie}
\title{famiglie}
\format{
Una data.table con i campi seguenti:
\describe{
  \item{\code{CMN}}{ Codice ISTAT del Comune (si veda la tabella \code{comuni} nel pacchetto \code{masteRgeo} ) }
  \item{\code{dominio}}{ \code{F}-Famiglie, \code{PF}-Popolazione Residente in Famiglia }
  \item{\code{anno}}{ Anno di riferimento }
  \item{\code{valore}}{ Valore del Dato }
}
}
\usage{
famiglie
}
\description{
Tabella contenente il conteggio delle Famiglie e della Popolazione Residente in Famiglia (nessun raggruppamento)
}
\details{
La tabella è indicizzata secondo i campi \code{CMN} ed \code{anno}.
}
\keyword{datasets}
