% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dati.R
\docType{data}
\name{istruzione_eta}
\alias{istruzione_eta}
\title{istruzione_eta}
\format{
Una data.table con i campi seguenti:
\describe{
  \item{\code{CMN}}{ Codice ISTAT del Comune (si veda la tabella \code{comuni} nel pacchetto \code{masteRgeo} ) }
  \item{\code{sesso}}{ \code{M}-Maschio, \code{F}-Femmina, \code{T}-Totale }
  \item{\code{eta}}{ Classi di eta. \code{Y09-24}: da 9 a 24 anni, \code{Y25-49}: da 25 a 49 anni, \code{Y50-64}: da 50 a 64 anni, \code{YGE65}: oltre 64 anni }
  \item{\code{istruzione}}{ Livello di Istruzione (si veda la tabella \code{cens_diz} per la descrizione delle etichette) }
  \item{\code{anno}}{ Anno di riferimento }
  \item{\code{valore}}{ Valore del Dato }
}
}
\usage{
istruzione_eta
}
\description{
Tabella contenente il conteggio della Popolazione Residente maggiore di anni 8 raggruppata per Sesso, Eta e Livello di Istruzione.
}
\details{
La tabella è indicizzata secondo i campi \code{CMN}, \code{sesso}, \code{eta}, \code{istruzione}, \code{anno}.
}
\keyword{datasets}
