% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masteRcensus.R
\docType{package}
\name{masteRcensus}
\alias{masteRcensus}
\title{masteRcensus: Pacchetto $R$ contenente Dati Territoriali dei Censimenti Permanenti Italia 2018-2020}
\description{
Le tabelle contenute nel pacchetto sono le seguenti:
- `diz_dati` Elenco degli attributi presenti nelle varie tabelle, con annesse etichette
- `residenti` 
- `famiglie` 
- `eta05` 
- `eta10` 
- `minori` 
- `stranieri_continenti` 
- `stranieri_aree` 
- `stranieri_paesi` 
- `istruzione` 
- `istruzione_eta` 
- `professione` 
- `professione_eta` 
- `pendolari`
}
\details{
Per riferimento, si consultino anche le seguenti tabelle contenute nel pacchetto \code{masteRgeo}:
 - `tipi_zone` elenco delle varie tipologie di *zone* geografiche ed amministrative italiane 
 - `zone` elenco completo delle *zone* italiane di ogni tipo
 - `comuni` elenco completo dei comuni italiani, con annesse gerarchie superiori.
}
