% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dati.R
\docType{data}
\name{professione}
\alias{professione}
\title{professione}
\format{
Una data.table con i campi seguenti:
\describe{
  \item{\code{CMN}}{ Codice ISTAT del Comune (si veda la tabella \code{comuni} nel pacchetto \code{masteRgeo} ) }
  \item{\code{sesso}}{ \code{M}-Maschio, \code{F}-Femmina, \code{T}-Totale }
  \item{\code{cittadinanza}}{ \code{I}-Italiano, \code{S}-Straniero, \code{T}-Totale }
  \item{\code{professione}}{ Condizione Professionale (si veda la tabella \code{cens_diz} per la descrizione delle etichette) }
  \item{\code{anno}}{ Anno di riferimento }
  \item{\code{valore}}{ Valore del Dato }
}
}
\usage{
professione
}
\description{
Tabella contenente il conteggio della Popolazione Residente maggiore di anni 14 raggruppata per Sesso, Cittadinanza e Condizione Professionale.
}
\details{
La tabella è indicizzata secondo i campi \code{CMN}, \code{sesso}, \code{cittadinanza}, \code{professione}, \code{anno}.
}
\keyword{datasets}
