% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dati.R
\docType{data}
\name{residenti}
\alias{residenti}
\title{residenti}
\format{
Una data.table con i campi seguenti:
\describe{
  \item{\code{CMN}}{ Codice ISTAT del Comune (vedere tabella \code{comuni} nel pacchetto \code{masteRgeo} ) }
  \item{\code{sesso}}{ \code{M}-Maschio, \code{F}-Femmina, \code{T}-Totale }
  \item{\code{cittadinanza}}{ \code{I}-Italiano, \code{S}-Straniero, \code{T}-Totale }
  \item{\code{anno}}{ Anno di riferimento }
  \item{\code{valore}}{ Valore del Dato }
}
}
\usage{
residenti
}
\description{
Tabella contenente il conteggio della Popolazione Residente raggruppata per Sesso e cittadinanza
}
\details{
La tabella è indicizzata secondo i campi \code{CMN}, \code{sesso}, \code{cittadinanza}, \code{anno}.
}
\keyword{datasets}
