% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dati.R
\docType{data}
\name{stranieri_continenti}
\alias{stranieri_continenti}
\title{stranieri_continenti}
\format{
Una data.table con i campi seguenti:
\describe{
  \item{\code{CMN}}{ Codice ISTAT del Comune (si veda la tabella \code{comuni} nel pacchetto \code{masteRgeo} ) }
  \item{\code{sesso}}{ \code{M}-Maschio, \code{F}-Femmina, \code{T}-Totale }
  \item{\code{provenienza}}{ Continente di provenienza (si veda la tabella \code{cens_diz} per la descrizione delle etichette) }
  \item{\code{anno}}{ Anno di riferimento }
  \item{\code{valore}}{ Valore del Dato }
}
}
\usage{
stranieri_continenti
}
\description{
Tabella contenente il conteggio degli Stranieri Residenti raggruppati per Sesso e Continente di provenienza
}
\details{
La tabella è indicizzata secondo i campi \code{CMN}, \code{sesso}, \code{provenienza}, \code{anno}.
}
\keyword{datasets}
